// core/_domain/enum/profile-component-repository.enum.ts
export enum PROFILE_COMPONENT_TYPE {
  PHONE = "PHONE",
  EMAIL = "EMAIL",
  IMAGE = "IMAGE",
  TEXT = "TEXT",
  HEADING = "HEADING",
  PARAGRAPH = "PARAGRAPH",
  LINK = "LINK",
  VIDEO = "VIDEO",
  MAP = "MAP",
  MICROSOFT_WORD = "MICROSOFT_WORD",
  PDF = "PDF",
  FACEBOOK = "FACEBOOK",
  TWITTER = "TWITTER",
  INSTAGRAM = "INSTAGRAM",
  LINKEDIN = "LINKEDIN",
  TIKTOK = "TIKTOK",
  FACEBOOK_POST = "FACEBOOK_POST",
  TWITTER_POST = "TWITTER_POST",
  INSTAGRAM_POST = "INSTAGRAM_POST",
  LINKEDIN_POST = "LINKEDIN_POST",
  TIKTOK_POST = "TIKTOK_POST",
  YOUTUBE_POST = "YOUTUBE_POST",
  SNAPCHAT = "SNAPCHAT",
  PINTEREST = "PINTEREST",
  DISCORD = "DISCORD",
  SLACK = "SLACK",
  YOUTUBE = "YOUTUBE",
  TWITCH = "TWITCH",
  GITHUB = "GITHUB",
  TELEGRAM = "TELEGRAM",
  WHATSAPP = "WHATSAPP",
  SPOTIFY = "SPOTIFY",
  APPLE_MUSIC = "APPLE_MUSIC",
  AMAZON_MUSIC = "AMAZON_MUSIC",
  DRIBBBLE = "DRIBBBLE",
  BEHANCE = "BEHANCE",
  PAYPAL = "PAYPAL",
  STRIPE = "STRIPE",
  AMAZON_PAY = "AMAZON_PAY",
  APPLE_PAY = "APPLE_PAY",
  GOOGLE_MEET = "GOOGLE_MEET",
  MICROSOFT_TEAMS = "MICROSOFT_TEAMS",
  INFO = "INFO",
  ZOOM = "ZOOM",
}

export enum PROFILE_COMPONENT_CATEGORY {
  MAIL = "MAIL",
  LINK = "LINK",
  VIDEO = "VIDEO",
  MAP = "MAP",
  FILE = "FILE",
  PHONE = "PHONE",
  PAYMENT = "PAYMENT",
  MUSIC = "MUSIC",
  IMAGE = "IMG",
  TEXT = "TEXT",
  SOCIAL_EMBED = "SOCIAL_EMBED",
}